Update app.py
Updated 'add_playlist_to_db'. Originally, it would just pull the playlist info. Now it also pulls song info as well and moves it to the song table. 

This works because we are able to see the amount of songs in a playlist when we pull a playlist. Meaning that the songs are available to get pulled. (There is a chance we could get the song-features from this like "Dancibility" etc. )